.github/pyproject.toml
[project]
name = "publisher-workflow-tools"
version = "0.1.0"
description = "Workflow & tools for PDF Publishing"
dependencies = [
  "pyyaml>=6.0",
  "pandas",
  "matplotlib",
  "numpy",
  "pillow"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tools"]
package-dir = {"" = "."}
